Correct QCD rate in datacard
The code was using the yield of the data instead of the yield of just QCD. This change takes the qcd yield from the bias template (and also updates the rate to 2-decimal precision so that it's the same as the other background processes). Also comments out the QCD norm systematic, which seemed to be causing a large amount of bias.